% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{lio_platforms}
\alias{lio_platforms}
\title{Platforms}
\usage{
lio_platforms(key = NULL, ...)
}
\arguments{
\item{key}{your Libraries.io API key. See `Auth` section in \pkg{rbraries}
manual file. optional}

\item{...}{curl options passed on to \code{\link[crul]{HttpClient}}}
}
\description{
Platforms
}
\examples{
\dontrun{
lio_platforms()
}
}
