% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user}
\alias{user}
\alias{user_template_fun}
\alias{lio_user}
\alias{lio_user_repositories}
\alias{lio_user_pkgs}
\alias{lio_user_pkg_contributions}
\alias{lio_user_repo_contributions}
\alias{lio_user_dependencies}
\title{User}
\usage{
user_template_fun(route)

lio_user(login, key = NULL, ...)

lio_user_repositories(login, key = NULL, ...)

lio_user_pkgs(login, key = NULL, ...)

lio_user_pkg_contributions(login, key = NULL, ...)

lio_user_repo_contributions(login, key = NULL, ...)

lio_user_dependencies(login, key = NULL, ...)
}
\arguments{
\item{login}{(character) a github login name. required}

\item{key}{your Libraries.io API key. See `Auth` section in \pkg{rbraries}
manual file. optional}

\item{...}{curl options passed on to \code{\link[crul]{HttpClient}}}
}
\description{
User
}
\examples{
\dontrun{
# user info
lio_user('ropensci')

# repositories
lio_user_repositories('ropensci')

# packages/projects
lio_user_pkgs('ropensci')

# package/project contributions
lio_user_pkg_contributions('ropensci')

# contributions
lio_user_repo_contributions('ropensci')

# dependencies
lio_user_dependencies('ropensci')
}
}
